+++ dislog/export_test.go
@@ -1,5 +1,5 @@
 package dislog
 
-// Split exposes split to the external test package, which is where the rest
+// Short exposes short to the external test package, which is where the rest
 // of the tests live.
-var Split = split
+var Short = short